utils: Extract command streaming process state
The command utility now separates event models from the runner, but the
runner still owns POSIX spawn setup and the streaming process state
machine.

That leaves child polling, selector registration, stdin writes, output
draining, and early-close termination in the same module as command startup
and completed-process capture.

This commit moves the spawned-process wrapper and StreamingProcess state
machine into utils.command_streaming while keeping spawn setup and
run_command wrappers in utils.command.

Subsequent commits will update architecture checks to guard the streaming
process boundary.
